creando providers (farm, user) y modificando objetos (user, farm)

+++ fronted_hackathon/hackathon_app/lib/ui/widgets/farm_card.dart
@@ -7,7 +7,7 @@ class FarmCard extends StatelessWidget {
   String farmName;
   int groups;
   String location;
-  double hectares;
+  int hectares;
   FarmCard(
       {super.key,
       required this.idFarm,
